requirements.txt
dash == 1.19.0
dataset == 1.4.5
discord.py ~= 1.7.0
pandas ~= 1.2.3
python-dotenv ~= 0.17.0
plotly ~= 4.14.3
SQLAlchemy ~= 1.4.5
